// src/api/mededu/models/TrainingInput.ts
export interface TrainingInput {
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    title: string;
    /**
     * 
     * @type {TrainingType}
     * @memberof TrainingInput
     */
    type: TrainingType;
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    department: string;
    /**
     * 
     * @type {Date}
     * @memberof TrainingInput
     */
    startAt: Date;
    /**
     * 
     * @type {number}
     * @memberof TrainingInput
     */
    capacity: number;
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    lecturer: string;
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    location?: string;
    /**
     * Online link skolenia; prazdny retazec znamena, ze skolenie je prezencne.
     * @type {string}
     * @memberof TrainingInput
     */
    onlineLink?: string;
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof TrainingInput
     */
    requirements?: string;
    /**
     * 
     * @type {TrainingStatus}
     * @memberof TrainingInput
     */
    status: TrainingStatus;
}