// src/api/mededu/models/TrainingStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MedEdu API
 * API pre spravu katalogu internych skoleni nemocnicneho personalu.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xcervenkak@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TrainingStatus = {
    Planned: 'planned',
    Cancelled: 'cancelled',
    Archived: 'archived'
} as const;
export type TrainingStatus = typeof TrainingStatus[keyof typeof TrainingStatus];


export function TrainingStatusFromJSON(json: any): TrainingStatus {
    return TrainingStatusFromJSONTyped(json, false);
}

export function TrainingStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TrainingStatus {
    return json as TrainingStatus;
}

export function TrainingStatusToJSON(value?: TrainingStatus | null): any {
    return value as any;
}