// src/api/mededu/models/TrainingType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MedEdu API
 * API pre spravu katalogu internych skoleni nemocnicneho personalu.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xcervenkak@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TrainingType = {
    Mandatory: 'mandatory',
    Department: 'department',
    Specialization: 'specialization',
    Online: 'online'
} as const;
export type TrainingType = typeof TrainingType[keyof typeof TrainingType];


export function TrainingTypeFromJSON(json: any): TrainingType {
    return TrainingTypeFromJSONTyped(json, false);
}

export function TrainingTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): TrainingType {
    return json as TrainingType;
}

export function TrainingTypeToJSON(value?: TrainingType | null): any {
    return value as any;
}